Check if nginx.conf exist This fixes #1

+++ wpinstall.sh
@@ -60,6 +60,12 @@ sudo chown $(whoami) /etc/hosts
 sudo sed -i -e "\$a127.0.0.1\t$domain_name" /etc/hosts
 sudo chown root /etc/hosts
 
+# create nginx configuration for $domain_name
+# check if nginx.conf exist
+if [ ! -f nginx.conf ]; then 
+    echo "nginx.conf file is missing in current directory. Aborting..." 1>&2
+    exit 1
+fi
 sudo cp nginx.conf /etc/nginx/sites-available/$domain_name
 sudo sed -i "s/domain_name/$domain_name/g" /etc/nginx/sites-available/$domain_name
 sudo ln -s /etc/nginx/sites-available/$domain_name /etc/nginx/sites-enabled/
